Deprecate src/Principal.php
