+++ tests/ArraysTest.php
@@ -23,7 +23,6 @@ use PHPUnit\Framework\TestCase;
 use Pipeline\Standard;
 
 /**
- * @covers \Pipeline\Principal
  * @covers \Pipeline\Standard
  *
  * @internal
@@ -75,7 +74,7 @@ final class ArraysTest extends TestCase
         $pipeline = new Standard();
 
         $reflectionClass = new \ReflectionObject($pipeline);
-        $reflectionProperty = $reflectionClass->getParentClass()->getProperty('pipeline');
+        $reflectionProperty = $reflectionClass->getProperty('pipeline');
         $reflectionProperty->setAccessible(true);
         $reflectionProperty->setValue($pipeline, [
             1 => 1,
