+++ tests/EdgeCasesTest.php
@@ -24,7 +24,6 @@ use function Pipeline\map;
 use Pipeline\Standard;
 
 /**
- * @covers \Pipeline\Principal
  * @covers \Pipeline\Standard
  *
  * @internal
